// packages/datadog-api-client-v1/models/MonitorOptions.d.ts
export declare class MonitorOptions {
    'aggregation'?: MonitorOptionsAggregation;
    /**
    * IDs of the device the Synthetics monitor is running on.
    */
    'deviceIds'?: Array<MonitorDeviceID>;
    /**
    * Whether or not to send a log sample when the log monitor triggers.
    */
    'enableLogsSample'?: boolean;
    /**
    * We recommend using the [is_renotify](https://docs.datadoghq.com/monitors/notifications/?tab=is_alert#renotify), block in the original message instead. A message to include with a re-notification. Supports the `@username` notification we allow elsewhere. Not applicable if `renotify_interval` is `None`.
    */
    'escalationMessage'?: string;
    /**
    * Time (in seconds) to delay evaluation, as a non-negative integer. For example, if the value is set to `300` (5min), the timeframe is set to `last_5m` and the time is 7:00, the monitor evaluates data from 6:50 to 6:55. This is useful for AWS CloudWatch and other backfilled metrics to ensure the monitor always has data during evaluation.
    */
    'evaluationDelay'?: number;
    /**
    * Whether the log alert monitor triggers a single alert or multiple alerts when any group breaches a threshold.
    */
    'groupbySimpleMonitor'?: boolean;
    /**
    * A Boolean indicating whether notifications from this monitor automatically inserts its triggering tags into the title.  **Examples** - If `True`, `[Triggered on {host:h1}] Monitor Title` - If `False`, `[Triggered] Monitor Title`
    */
    'includeTags'?: boolean;
    /**
    * Whether or not the monitor is locked (only editable by creator and admins).
    */
    'locked'?: boolean;
    /**
    * How long the test should be in failure before alerting (integer, number of seconds, max 7200).
    */
    'minFailureDuration'?: number;
    /**
    * The minimum number of locations in failure at the same time during at least one moment in the `min_failure_duration` period (`min_location_failed` and `min_failure_duration` are part of the advanced alerting rules - integer, >= 1).
    */
    'minLocationFailed'?: number;
    /**
    * Time (in seconds) to allow a host to boot and applications to fully start before starting the evaluation of monitor results. Should be a non negative integer.
    */
    'newHostDelay'?: number;
    /**
    * The number of minutes before a monitor notifies after data stops reporting. Datadog recommends at least 2x the monitor timeframe for metric alerts or 2 minutes for service checks. If omitted, 2x the evaluation timeframe is used for metric alerts, and 24 hours is used for service checks.
    */
    'noDataTimeframe'?: number;
    /**
    * A Boolean indicating whether tagged users is notified on changes to this monitor.
    */
    'notifyAudit'?: boolean;
    /**
    * A Boolean indicating whether this monitor notifies when data stops reporting.
    */
    'notifyNoData'?: boolean;
    /**
    * The number of minutes after the last notification before a monitor re-notifies on the current status. It only re-notifies if it’s not resolved.
    */
    'renotifyInterval'?: number;
    /**
    * A Boolean indicating whether this monitor needs a full window of data before it’s evaluated. We highly recommend you set this to `false` for sparse metrics, otherwise some evaluations are skipped. Default is false.
    */
    'requireFullWindow'?: boolean;
    /**
    * Information about the downtime applied to the monitor.
    */
    'silenced'?: {
        [key: string]: number;
    };
    /**
    * ID of the corresponding Synthetic check.
    */
    'syntheticsCheckId'?: string;
    'thresholdWindows'?: MonitorThresholdWindowOptions;
    'thresholds'?: MonitorThresholds;
    /**
    * The number of hours of the monitor not reporting data before it automatically resolves from a triggered state.
    */
    'timeoutH'?: number;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): MonitorOptions;
    static serialize(data: MonitorOptions): {
        [key: string]: any;
    };
    constructor();
}